api/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1